student_handle: fix import for student_handle [1/3]
